transfer files to new location and modify documentation
