--- skfda/exploratory/fpca/__init__.py
@@ -1,1 +1,0 @@
-from ._fpca import FPCABasis, FPCADiscretized
