+++ skfda/preprocessing/dim_reduction/__init__.py
